Añadido nombres de montaje en combobox
